[Frontend][Driver] Remove `compile_from_mlir`. (#1181)
**Context:** In order to make the compiler driver a standalone
executable, we would to simplify the compiler driver interface. I.e.,
the compiler driver executable should just produce a shared object. The
`struct FunctionAttributes` which contained two string fields (one for
the name of the entry point of the function and the other for the return
values of the entry point) was only used when lowering MLIR from the
textual representation, which has been useful for debugging.

**Description of the Change:** Remove `compile_from_mlir`.

**Benefits:** The benefit is that the compiler driver can just focus on
compiling without having to return extra metadata which simplifies the
interface between the compiler driver and the python process running it.

**Possible Drawbacks:** None, use `replace_ir` instead.

```