Implement common Catalyst callable base class
The main purpose is to identify callable objects from Catalyst
transforms, which modify the behaviour of a given user function.
For those callables, we want to be able to apply additional wrappers to
the user function without making changes to the Catalyst callable
instance. One application of this is autograph AST transforms. We need
to transform the wrapped user function but not the Catalyst callable
code itself.