Extend the cancel_inverses pass to cancel adjoints of an arbitrary unitary (#1186)
**Context:**
Extend the cancel_inverses pass to cancel adjoints of an arbitrary
unitary.

**Description of the Change:**
- Add an analysis to check whether a quantum gate and its parent gate
are well-matched for the purposes of peephole optimizations like
cancel_inverse and merge_rotation. See `mlir/lib/Quantum/Transforms
/VerifyParentGateAnalysis.hpp`.
- Include a new cancel_inverse pattern "---U---U{adj}---" for arbitrary
U.

**Benefits:**
Arbitrary inverses can now be cancelled, in addition to the named
Hermitian gates.

**Possible Drawbacks:**
Two potential improvements:
1. Control wires' values don't have to match one-to-one exactly, but
only the set of true wires and the set of false wires need to match, and
within the set the wires can be of arbitrary order. The exact one-to-one
being checked right now is overly conservative.
2. Instead of manually adding the pattern for `customOp`, `unitaryOp`
and `multiRZOp` separately, we might want to use the [quantum op
interface](https://github.com/PennyLaneAI/catalyst/blob/main/mlir/include/Quantum/IR/QuantumInterfaces.td).
However, mlir pattern matcher expects the op being matched to be
concrete (so that it can look for a name in the textual IR) instead of
being an abstract interface. Some redesign of the interface and its
usage might be beneficial.

**Related GitHub Issues:**
[sc-74518]