[MLIR] Emit error instead of segfault in current scatter lowering (#1217)
**Context:** The current scatter lowering makes some assumptions and
only handles a subset of the inputs. However, it is too optimistic and
will attempt to convert inputs that it can't handle. In particular if
there are more than one input update and result tensors, it will
segfault. This condition may never happen but it is valid IR. It may
never happen if another pass runs before that canonicalizes
mhlo.scatter. However, it is good to emit an error in case the
canonicalization was not run before.

**Description of the Change:** Emit error instead of segfaulting

**Benefits:** No segfaults

**Possible Drawbacks:** None

**Related GitHub Issues:**

Depends on #1216