Do not mandate order of outlined tape functions in split tape lit test (#1228)
There is a
[test](https://github.com/PennyLaneAI/catalyst/blob/main/mlir/test/Quantum/SplitMultipleTapesTest.mlir#L67)
in the split tape lit tests where there are two multi-tape functions in
a module, and we split both of them.

It is unclear why, but on CI the outlined single-tape functions are,
very occasionally (about 4 times in 3 months), processed and/or outlined
in the reversed order. This behavior was unable to be reproduced over
12*2e9 local runs.

#1130 was an attempt to fix this, which did reduce the flaky frequency
by a lot; however there is a small remaining mysterious flakiness.

Recognizing that the outlined single-tape functions' order does not
actually matter, we change the `CHECK`s to `CHECK-DAG`s, which do not
mandate the order of the checked string during filecheck.