Apply clang-tidy to lightning.qubit runtime  (#1261)
**Context:**
Catalyst should start conforming to `clang-tidy`. As a first step, we
apply it to the files in the lightning runtime.

Note that this came about both as something we should do anyways, and as
a concrete blocker to downstream lightning's catalyst interface to
lightning repo (where clang tidy warnings as errors is enabled). We
would like to fix all errors reported
[here](https://github.com/PennyLaneAI/pennylane-lightning/actions/runs/11609135273/job/32325785221?pr=960).

**Description of the Change:**
1. Update `runtime/.clang-tidy` and clang-tidy version in `requirements.txt` to
match those in the Lightning repo
2. Apply clang-tidy to lightning.qubit runtime. 
The target files are those in
`runtime/lib/backend/lightning/lightning_dynamic`, alongside some common
files in the backend as well.

**Benefits:**
Better adherence to clang rules.

[sc-77220]