Replace pybind11 with nanobind in frontend (#1173)
**Context:**

The Catalyst frontend contains a small Python extension module at
[frontend/catalyst/utils/wrapper.cpp](https://github.com/PennyLaneAI/catalyst/blob/main/frontend/catalyst/utils/wrapper.cpp),
which is an importable Python module written in C/C++ used to wrap the
entry point of compiled programs. The Python-C++ bindings were
originally implemented using pybind11. This PR is part of a larger
effort to replace all pybind11 code with nanobind.

**Description of the Change:**

This change replaces all the pybind11 code in the frontend with the
equivalent nanobind objects and operations.

It was also necessary to modify the frontend build system to build the
`wrapper` module using CMake for compatibility with nanobind, rather
than in `setup.py` with the `intree_extensions` setuptools utility
included with pybind11.

**Benefits:**

See Epic [68472](https://app.shortcut.com/xanaduai/epic/68472) for a
list of nanobind's benefits.

-----


[[sc-72837](https://app.shortcut.com/xanaduai/story/72837/replace-pybind11-with-nanobind-in-the-frontend)]

---------

Co-authored-by: Lee James O'Riordan <mlxd@users.noreply.github.com>