Added an empty pass boilerplate as a tutorial (#872)
**Context:** Added a tutorial for writing an empty pass boilerplate in
the `catalyst` dialect that prints out hello world. The boilerplate code
is on branch `tutorial/my-hello-world-pass` in commit
https://github.com/PennyLaneAI/catalyst/commit/ba7b3438667963b307c07440acd6d7082f1960f3

Also fixed the errors in the pattern match and rewrite tutorial code,
and created a complete code listing for that tutorial. It is on branch 
`tutorial/merge-unitary-gates-pass` in commit
https://github.com/PennyLaneAI/catalyst/commit/9afcc3500e12e5a51b78dda76cd4d27bdf4c8905

**Benefits:** A tutorial for the pass boilerplate infrastructure for
those who want to write custom passes.


[sc-67240]