Catalyst support for `qml.CosineWindow` (#1166)
Compiling a circuit with QJIT that begins with a CosineWindow state
preparation results in a compiler error. This is because Catalyst has
optimizations for when the first operation is a state preparation, but
these optimizations are not supported for CosineWindow. A workaround is
to only perform the state preparation optimizations on the `StatePrep`
and `BasisState` classes, rather than the more general `StatePrepBase`
class.

The latest version of this PR makes the PRs to PennyLane [(PR
#6318)](https://github.com/PennyLaneAI/pennylane/pull/6318) and
PennyLane lightning [(PR
#929)](https://github.com/PennyLaneAI/pennylane-lightning/pull/929)
unnecessary.

---------

Co-authored-by: Ali Asadi <10773383+maliasadi@users.noreply.github.com>
Co-authored-by: David Ittah <dime10@users.noreply.github.com>
Co-authored-by: erick-xanadu <110487834+erick-xanadu@users.noreply.github.com>