[WIP] replace new compiler driver by calling compiler-cli using subprocess (#1285)
**Context:** The old compiler driver api which uses the pybind11 as an
interface between the python frontend and the compiler should be
replaced with the new CLI tool (catalyst-cli).

**Description of the Change:**
Removes the pybind11 interface in favour of calling catalyst-cli using
python subprocess.

**Benefits:**
the catalyst-cli will be completely independent of the python frontend
and consequently, compiler symbols will be separated from the frontend

**Possible Drawbacks:**
Using subprocess requires us to connect with the catalyst-cli using
files on disk. Therefore we need to save the IR to a file temporarily
and parse it again in the compiler which can be problematic when IR size
grows. This should be improved in future using interprocess
communication.

[sc-73537]
[sc-73538]