Fix test matrix configuration
Sometimes `pip install --upgrade` is not sufficient to have pip install a desired version of a package that is already installed, in particular when the requested version is lower or equal to the installed one. In this case, `--force` (or `--force-reinstall`) is required. To avoid breaking existing version pins, `--no-deps` is also required.