Replace pybind11 with nanobind in Catalyst Runtime (#1293)
**Context:**

The Catalyst runtime makes use of Pybind11 in two places:

* The callback registry:
https://github.com/PennyLaneAI/catalyst/tree/main/runtime/lib/registry
* The OpenQASM device backends:
https://github.com/PennyLaneAI/catalyst/tree/main/runtime/lib/backend/openqasm

This PR is part of a larger effort to replace all pybind11 code with
nanobind.

**Description of the Change:**

This change replaces all the pybind11 code in
[runtime/lib/registry/Registry.cpp](https://github.com/PennyLaneAI/catalyst/blob/main/runtime/lib/registry/Registry.cpp)
and in
[runtime/lib/backend/openqasm/openqasm_python_module.cpp](https://github.com/PennyLaneAI/catalyst/blob/main/runtime/lib/backend/openqasm/openqasm_python_module.cpp)
with the equivalent nanobind objects and operations.

**Benefits:**

See Epic [68472](https://app.shortcut.com/xanaduai/epic/68472) for a
list of nanobind's benefits.

-----


[[sc-72877](https://app.shortcut.com/xanaduai/story/72877/replace-pybind11-with-nanobind-in-the-runtime)]