Source OpenBLAS from dedicated package (#1322)
SciPy now publishes a standalone OpenBLAS distribution on PyPI
`scipy-openblas32`, which will allow us to unpin scipy in `setup.py`
(the pin was required since recent versions of scipy, `1.14` and up, no
longer distribute OpenBLAS in their wheels on macOS).

Symbols in this library are prefixed with `scipy_` which requires
adjusting our symbol names. Library discovery is now made easier since
it's uniform for linux and mac. However, some versions of this package
need to be excluded due to precision and linking issues.

[sc-78888]

---------

Co-authored-by: Joey Carter <joseph.carter@xanadu.ai>