[MLIR] load and execute plugins (#1287)
**Context:** After #1285 we will be able to compile LLVM and our MLIR
tools with default visibility. Default visibility is needed for MLIR
plugins.

**Description of the Change:** This PR adds the ability to load and
execute plugins from catalyst-cli and quantum-opt.

**Benefits:** Users can build the standalone module easily with `make
standalone-plugin` and can run it with `catalyst-cli`.

**Possible Drawbacks:** Something I don't like about this PR is the need
to copy the standalone folder from LLVM/MLIR. However, I do make some
changes such that the plugin test tests that it works with catalyst-cli.
See
[here](https://github.com/PennyLaneAI/catalyst/pull/1287/files#diff-67e240bcc3a86bd66928e3454c1922126a2ec38406b5ddc26d2138bd33024850R51-R61)
for an example.

**Related GitHub Issues:**

[sc-73562]
[sc-73565]
[sc-73566]

Before merging, we need to clear the cache of all LLVM, MHLO, Enzyme as
they need to be rebuilt with VISIBILITY=default. Otherwise, trying to
build the standalone plugin will fail.

---------

Co-authored-by: David Ittah <dime10@users.noreply.github.com>