Update Catalyst to use the toml module in PennyLane (#1275)
**Context:**

This PR is part of the new device capabilities initiative to improve
feature parity between PennyLane and Catalyst. See [this
ADR](https://github.com/PennyLaneAI/adrs/pull/78) for more context. A
new TOML schema has been defined, and the relevant module implemented in
PennyLane:
- https://github.com/PennyLaneAI/pennylane/pull/6407
- https://github.com/PennyLaneAI/pennylane/pull/6433.

**Description of the Change:**

- Removes the `toml` module and replaces all usages with imports from
PennyLane
- Update usages of `DeviceCapabilities` to match the new definition in
https://github.com/PennyLaneAI/pennylane/pull/6407/
- Removes use of about-to-be-deprecated return_type of MPs
- Updates all TOML files to use the new schema.
- A PR opened in Lightning
https://github.com/PennyLaneAI/pennylane-lightning/pull/988 that updates
the TOML file of the Lightning devices to the new schema.

**Benefits:**

A step towards feature parity between PennyLane and Catalyst.

**Possible Drawbacks:**

- Per discussions when the ADR was developed, `operator.gates.decomp`
and `operator.gates.matrix` are removed, and the TOML file no longer
prescribes to the framework how an operator should be handled. This will
change the behaviour of decomposition for some circuits.

**Related GitHub Issues:**

[sc-71888]
[sc-71728]
[sc-71730]

---------

Co-authored-by: Romain Moyard <rmoyard@gmail.com>