Use LAPACKE interface for Accelerate on macOS (#1328)
Final update to the LAPACK sourcing issue.

On arm64 macOS systems, we were observing numerical disagreements when
using OpenBLAS provided by SciPy. This PR adds our build of the LAPACKE
C-interface for LAPACK, which redirects to the Accelerate framework
provided by Apple.

On Linux and x86 macOS systems, we continue to use OpenBLAS. x86 macOS
is included there because we observed similar numerical issues with
LAPACKE+Accelerate on this platform than with SciPy OpenBLAS on arm64
macOS.


[sc-78888]

---------

Co-authored-by: Joey Carter <joseph.carter@xanadu.ai>