[Frontend|MLIR] Each qnode has its own schedule (#1323)
**Context:** Instead of relying on function names, passes that can be
scheduled on qnodes can now act on the whole qnode module. Each qnode
now comes with its own schedule, which will allow for **possibly**
multithreaded compilation of each qnode according to a schedule.

**Description of the Change:** the `transform_named_sequence_p`
primitive is now inside the qnode primitive. Each qnode has one of
these. The `apply_registered_pass_p` modifies the correct
`transform_named_sequence_p` during the lowering.

**Benefits:** qnode specific schedules

**Possible Drawbacks:** Maybe we can have the transform be an attribute
instead of an instruction on the qnode (at the JAXPR level). I find it a
bit odd how this is modeled, but it is working.

**Related GitHub Issues:**

---------

Co-authored-by: Raul Torres <138264735+rauletorresc@users.noreply.github.com>