[Frontend] Refactor registering passes and caching primitive mechanisms (#1349)
**Context:** Registering a pass to run in MLIR was a bit messy. This
code simplifies that process.

**Description of the Change:**

* Removes the `apply_registered_pass_p` primitive and instead denotes
the pass as an object. This object is used as a JAX static parameter and
its properties are used to lower the proper transform schedule.
* Simplify the process of adding a pass to the pipeline by just
appending it to a kwargs sequence.
* Simplify the process of mlir_quantum_decomposition for the ion pass
(@joeycarter). Ideally we should have a field in the device that
corresponds to the default passes to be applied. This is instead of
detecting the device during the lowering and adding a pass specific for
that device.
* Refactors the caching mechanism for primitives.

**Benefits:** Less complexity, more maintainable code. We can still
improve upon this. This change makes it easier to implement loading pass
plugins from python.

**Possible Drawbacks:** Changes some of the original formulation of the
pipeline (instead of having a dictionary, users can now just pass a list
of Pass objects)

**Related GitHub Issues:**