Add first draft of OQD trapped-ion device TOML databases and APIs (#1348)
**Context:**

PR #1260 will introduce an MLIR dialect (the _Ion Dialect_) for Open
Quantum Design (OQD) trapped-ion quantum devices and the means to lower
quantum-dialect MLIR to the ion dialect. The ion dialect requires data
specific to the target device in order to perform the lowering. For
example, the compiler requires data for the ions used in the device,
such as their species and energy levels, parameters determined from
calibration runs, such as laser pulse frequencies and Rabi frequencies,
and other parameters pertaining to the experimental apparatus. This PR
creates these databases and the APIs to load and interact with them.

**Description of the Change:**

Three database objects have been created, contained in the module
frontend/catalyst/third_party/oqd/oqd_database_managers.py:

1. `OQDDeviceDatabase`: for storing properties about the OQD device,
such as hardware specification and parameters relating to the
experimental apparatus.
2. `OQDQubitDatabase`: for storing properties about the physical qubits
used in the trapped-ion device. These are generally physical constants
of the ion species being used, such as their energy levels and
transitions, and phonon parameters.
3. `OQDBeamDatabase`: for storing parameters about the laser beams used
to interact with the qubits, such as the Rabi frequency, detuning, phase
and polarization parameters for the given energy-level transition.

These databases have been implemented using TOML configuration files for
the time being for simplicity and ease-of-use. In the future, some of
these parameters may be stored in a SQL database, particularly those
determined in calibration runs (these parameters will thus be updated
regularly). For the moment, we assume that there is the ability to
extract the relevant parameters from the database and dump them into
TOML format immediately before compilation. The database APIs listed
above have been separated into three objects for this reason—to separate
out parameters that are device constants (`OQDDeviceDatabase`), physical
constants (`OQDQubitDatabase`), and parameters that require calibration
and updating (`OQDBeamDatabase`).

The implementation details of the OQD trapped-ion device are still being
finalized, so the list of parameters and how they are organized is
subject to change. This PR creates a template database on which we can
build and start using to test drive the ion dialect.

These changes are largely self-contained and should not interfere with
regular Catalyst usage.

-----

This PR also introduces the `toml_utils` module to extend the
functionality of the TOML database files. It adds the `safe_eval`
function, which allows inputting values as arithmetic expressions. For
example, instead of writing

```toml
x = 6283185307.179586  # 2 * pi * 1e9
```

we can instead input this value directly as:

```toml
x = "2 * math.pi * 1e9"
```

The main advantage of this approach is that it uses self-documenting
values: it is clear what the origin of the value is without having to
keep the comment in sync with the value itself.

We define `safe_eval` to perform the expression evaluation rather than
use Python's built-in `eval()` function, since `eval`'s ability to
perform arbitrary code execution of a user's input makes it inherently
unsafe. The functionality of `safe_eval` is deliberately limited to
basic mathematical operations to prevent malicious code, intentional or
not, from being evaluated.

-----


[[sc-73706](https://app.shortcut.com/xanaduai/story/73706/create-a-list-of-hardware-info-calibration-data-necessary-for-the-compiler-0-5)],
[[sc-73707](https://app.shortcut.com/xanaduai/story/73707/add-hardware-info-to-the-toml-json-specs-1-5)]

---------

Co-authored-by: paul0403 <79805239+paul0403@users.noreply.github.com>