Remove shots attr from `expval` and `var` operations (#1360)
**Context:**
As part of the work to support dynamically shaped measurement
primitives, shots is now tied to the device through a SSA value, so it
can be dynamic. See #1310 .

Correspondingly, the static shots int literal attribute of the `expval`
and `var` operations in mlir should be removed. This is more so the case
since their shots values are not actually used in their respective
runtime CAPIs.

**Description of the Change:**
Remove shots attr from `expval` and `var` operations in mlir, and
frontend tracing and jax primitives.

**Benefits:**
Unification of how measurement operations handle shots.

[sc-78845]