Fix latest/latest/latest failure (#1377)
**Context:** Check CPL latest/latest/latest is failing because the
device toml file is not found when instantiating an OQD device. This
happened because Check CPL installs catalyst using `make catalyst`, not
`make all`, and the oqd build target was not added as a dependency of
`make catalyst`.

**Description of the Change:** Add `oqd` to dependencies of `make
catalyst`. Also, I've made an opportunistic fix to an incorrect GitHub
action title: "Download PennyLane-Lightning (**latest**)" where it
should be "Download PennyLane-Lightning (**release-candidate**)"