Fix OQC target by not relying on Development.Embed
The `pybind11_add_module` commands expects that Development.Embed is
available when used with the SHARED option. The default MODULE option
has no such requirement. For the wheel builds this poses a problem since
manylinux images don't include libpython.

Strangely enough, if the legacy Python discovery from Pybind11 is used
(by not calling find_package(Python) first), the command can still be
used with the SHARED option (which is how we got away with it so far).