Create a finite state machine analysis pass to propagate Pauli basis states, and use it to disentangle CNOTs and SWAPs (#1154)
**Context:**
In circuits that mostly involve the six Pauli eigenstates, the effect of
gates can be described by a simple six-state finite state machine. This
allows us to deduce their midcircuit state at compile time.

The knowledge of the snapshot midcircuit states can be useful when:
- Deducing CNOT/Toffoli/SWAP/... effect and decouple them. We decouple
CNOT and SWAP in this PR.
- Replace complicated circuits whose results are |01+-LR> with standard
preparation routines
  - Potentially more

**Description of the Change:**
Added an analysis for propagating simple states (and tests).
Also added CNOT and SWAP decomposition pass `disentangle-CNOT`,
`disentangle-SWAP` (and tests).
Credit to [Ritu Thombre](https://github.com/ritu-thombre99) for the work
on the SWAP gates!

**Benefits:**
More compile time circuit information for potential compile time
transformation/optimization.

**Possible Drawbacks:**
This algorithm is from an external paper. We need to properly cite it. 
https://arxiv.org/abs/2012.07711 (fig.5)

**Related GitHub Issues:** closes #1268 

[sc-74649]
[sc-74650]

---------

Co-authored-by: Ritu Thombre <renitath99@gmail.com>
Co-authored-by: Ritu Thombre <42207923+ritu-thombre99@users.noreply.github.com>