Restore braket tests (#1399)
**Context:**
Braket tests were removed temporarily because the dev version of
PennyLane needed for Catalyst tests was incompatible with the Braket
plugin in the interim between removing legacy operator arithmetic from
PennyLane and releasing a version of the Braket plugin with no
references to legacy opmath.

Braket has released, so we expect things to work again now. 

**Description of the Change:**
This PR reverses the changes disabling Braket tests in #1308 and #1321.

[sc-78774]