Change function to expect list instead of tuple.
We need a tuple in order for JAX primitive binding.
This is because a list can be modified and therefore have side effects.
A tuple cannot be modified so it can be bound to a parameter in a JAX primitive.

However, it is better from the developer perspective to have this parameter be a list.
It is easier conceptually to say that the only thing one is doing is appending an object to a list rather than creating a new tuple with a new object.