Update `from_plxpr` to use `PlxprInterpreter` base class (#1398)
**Context:**

When `from_plxpr` was added, we did not have any other plxpr
interpreters or frameworks for them. Now we have the `PlxprInterpreter`
class that handles a lot of the infrastructure for converting plxpr to
catalxpr.

**Description of the Change:**

Uses the `PlxprInterpreter` class to cut down on repeated code and
infrastructure in the `from_plxpr` implementation. This also makes it to
the interpreters follow the same pattern as other plxpr interpreters.

**Benefits:**

**Possible Drawbacks:**

**Related GitHub Issues:**

[sc-81469]

---------

Co-authored-by: David Ittah <dime10@users.noreply.github.com>